+++ simvue/run.py
@@ -1,4 +1,3 @@
-import configparser
 import datetime
 import hashlib
 import logging
@@ -6,16 +5,17 @@ import mimetypes
 import os
 import re
 import multiprocessing
+import pickle
 import socket
 import subprocess
 import sys
 import time as tm
 import platform
 import uuid
-import requests
 
 from .worker import Worker
 from .simvue import Simvue
+from .serialization import Serializer
 from .models import RunInput
 from .utilities import get_auth, get_expiry
 from pydantic import ValidationError
@@ -96,18 +96,25 @@ def get_system():
     return system
 
 
-def calculate_sha256(filename):
+def calculate_sha256(filename, is_file):
     """
     Calculate sha256 checksum of the specified file
     """
     sha256_hash = hashlib.sha256()
-    try:
-        with open(filename, "rb") as fd:
-            for byte_block in iter(lambda: fd.read(CHECKSUM_BLOCK_SIZE), b""):
-                sha256_hash.update(byte_block)
-            return sha256_hash.hexdigest()
-    except:
-        pass
+    if is_file:
+        try:
+            with open(filename, "rb") as fd:
+                for byte_block in iter(lambda: fd.read(CHECKSUM_BLOCK_SIZE), b""):
+                    sha256_hash.update(byte_block)
+                return sha256_hash.hexdigest()
+        except:
+            pass
+    else:
+        if isinstance(filename, str):
+            sha256_hash.update(bytes(filename, 'utf-8'))
+        else:
+            sha256_hash.update(bytes(filename))
+        return sha256_hash.hexdigest()
 
     return None
 
@@ -255,11 +262,11 @@ class Run(object):
 
         self._check_token()
 
-        # compare with pydantic RunInput model    
+        # compare with pydantic RunInput model
         try:
             runinput = RunInput(**data)
-        except ValidationError as e:
-            self._error(e)
+        except ValidationError as err:
+            self._error(err)
 
         self._simvue = Simvue(self._name, self._uuid, self._mode, self._suppress_errors)
         name = self._simvue.create_run(data)
@@ -474,9 +481,9 @@ class Run(object):
 
         return True
 
-    def save(self, filename, category, filetype=None, preserve_path=False):
+    def save(self, filename, category, filetype=None, preserve_path=False, name=None, allow_pickle=False):
         """
-        Upload file
+        Upload file or object
         """
         if self._mode == 'disabled':
             return True
@@ -489,12 +496,16 @@ class Run(object):
             self._error('Run is not active')
             return False
 
-        if not os.path.isfile(filename):
-            self._error(f"File {filename} does not exist")
-            return False
+        is_file = False
+        if isinstance(filename, str):
+            if not os.path.isfile(filename):
+                self._error(f"File {filename} does not exist")
+                return False
+            else:
+                is_file = True
 
         if filetype:
-            mimetypes_valid = []
+            mimetypes_valid = ['application/vnd.plotly.v1+json']
             mimetypes.init()
             for _, value in mimetypes.types_map.items():
                 mimetypes_valid.append(value)
@@ -508,24 +519,38 @@ class Run(object):
             data['name'] = filename
             if data['name'].startswith('./'):
                 data['name'] = data['name'][2:]
-        else:
+        elif is_file:
             data['name'] = os.path.basename(filename)
+
+        if name:
+            data['name'] = name
+
         data['run'] = self._name
         data['category'] = category
-        data['checksum'] = calculate_sha256(filename)
-        data['size'] = os.path.getsize(filename)
-        data['originalPath'] = os.path.abspath(os.path.expanduser(os.path.expandvars(filename)))
+
+        if is_file:
+            data['size'] = os.path.getsize(filename)
+            data['originalPath'] = os.path.abspath(os.path.expanduser(os.path.expandvars(filename)))
+            data['checksum'] = calculate_sha256(filename, is_file)
 
         # Determine mimetype
-        if not filetype:
+        mimetype = None
+        if not filetype and is_file:
             mimetypes.init()
             mimetype = mimetypes.guess_type(filename)[0]
             if not mimetype:
                 mimetype = 'application/octet-stream'
-        else:
+        elif is_file:
             mimetype = filetype
 
-        data['type'] = mimetype
+        if mimetype:
+            data['type'] = mimetype
+
+        if not is_file:
+            data['pickled'], data['type'] = Serializer().serialize(filename, allow_pickle)
+            data['checksum'] = calculate_sha256(data['pickled'], False)
+            data['originalPath'] = ''
+            data['size'] = sys.getsizeof(data['pickled'])
 
         # Register file
         if not self._simvue.save_file(data):
@@ -678,7 +703,8 @@ class Run(object):
                   notification='none',
                   pattern=None):
         """
-        Creates an alert with the specified name (if it doesn't exist) and applies it to the current run
+        Creates an alert with the specified name (if it doesn't exist)
+        and applies it to the current run
         """
         if self._mode == 'disabled':
             return True