+++ simvue/sender.py
@@ -12,6 +12,13 @@ from .utilities import get_offline_directory, create_file, remove_file
 
 logger = logging.getLogger(__name__)
 
+def update_name(name, data):
+    """
+    Update name in metrics/events
+    """
+    for item in data:
+        item['run'] = name
+
 def add_name(name, data, filename):
     """
     Update name in JSON
@@ -67,7 +74,11 @@ def sender():
         elif run.endswith('terminated'):
             status = 'terminated'
 
-        current = run.replace('/running', '').replace('/completed', '').replace('/failed', '').replace('/terminated', '').replace('/created', '')
+        current = run.replace('/running', '').\
+                  replace('/completed', '').\
+                  replace('/failed', '').\
+                  replace('/terminated', '').\
+                  replace('/created', '')
 
         if os.path.isfile("f{current}/sent"):
             if status == 'running':
@@ -147,37 +158,41 @@ def sender():
             # Handle metrics
             if '/metrics-' in record:
                 logger.info('Sending metrics for run %s', run_init['name'])
-                remote.send_metrics(msgpack.packb(get_json(record, name), use_bin_type=True))
+                data = get_json(record, name)
+                update_name(run_init['name'], data)
+                remote.send_metrics(msgpack.packb(data, use_bin_type=True))
                 rename = True
 
             # Handle events
-            if '/event-' in record:
-                logger.info('Sending event for run %s', run_init['name'])
-                remote.send_event(msgpack.packb(get_json(record, name), use_bin_type=True))
+            if '/events-' in record:
+                logger.info('Sending events for run %s', run_init['name'])
+                data = get_json(record, name)
+                update_name(run_init['name'], data)
+                remote.send_event(msgpack.packb(data, use_bin_type=True))
                 rename = True
 
             # Handle updates
             if '/update-' in record:
                 logger.info('Sending update for run %s', run_init['name'])
-                remote.update(get_json(record, name))
+                remote.update(get_json(record, name), run_init['name'])
                 rename = True
 
             # Handle folders
             if '/folder-' in record:
                 logger.info('Sending folder details for run %s', run_init['name'])
-                remote.set_folder_details(get_json(record, name))
+                remote.set_folder_details(get_json(record, name), run_init['name'])
                 rename = True
 
             # Handle alerts
             if '/alert-' in record:
                 logger.info('Sending alert details for run %s', run_init['name'])
-                remote.add_alert(get_json(record, name))
+                remote.add_alert(get_json(record, name), run_init['name'])
                 rename = True
 
             # Handle files
             if '/file-' in record:
                 logger.info('Saving file for run %s', run_init['name'])
-                remote.save_file(get_json(record, name))
+                remote.save_file(get_json(record, name), run_init['name'])
                 rename = True
 
             # Rename processed files