+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+## v0.8.0
+
+* Support NumPy arrays, PyTorch tensors, Matplotlib and Plotly plots and picklable Python objects as artifacts.
+* (Bug fix) Events in offline mode didn't work.
+
 ## v0.7.2
 
 * Pydantic model is used for input validation.